+++ banzai/tests/test_array_utils.py
@@ -6,6 +6,7 @@ from banzai.utils import array_utils
 
 pytestmark = pytest.mark.array_utils
 
+
 def test_pruning_nans():
     a = np.arange(100, dtype=float)
     b = np.arange(100, dtype=float) + 100
@@ -16,7 +17,7 @@ def test_pruning_nans():
     c[78] = np.nan
 
     t = Table([a, b, c], names=('a', 'b', 'c'))
-    t = array_utils.prune_nans_from_table(t)
+    t = t[~array_utils.find_nans_in_table(t)]
     assert len(t) == 97
     assert 51 not in t['a']
     assert 32 not in t['a']