+++ CHANGES.md
@@ -1,3 +1,25 @@
+1.17.0 (2023-04-24)
+-------------------
+- We now omit sources in the photometry stage that have an area larger than 1000 pixels as they lead to long 
+  processing times and are almost invariably spurious.
+
+1.16.1 (2023-04-23)
+-------------------
+- Correction to aperture photometry. We were incorrectly using the radius instead of the diameter
+
+1.16.0 (2023-04-18)
+-------------------
+- Calibration frames are now associated with output data products rather than frames
+  so that we have more than one calibration data product produced per frame.
+
+1.15.2 (2023-04-12)
+-------------------
+- Fix to fpacking data when the image data array is None
+
+1.15.1 (2024-02-29)
+-------------------
+- Minor fixes in photometry when there are bad pixels near the image edges
+
 1.15.0 (2024-02-14)
 -------------------
 - Migrated photometry extraction to be done by astropy's photutils instead of SEP.